# Backend/docs/swagger.yaml
definitions:
...
  controller.loginRequest:
    properties:
      email:
        example: admin@gmail.com
        type: string
      password:
        example: admin@1234
        type: string
    required:
    - email
    - password
    type: object
  controller.loginResponse:
    properties:
      data:
        properties:
          access_token:
            type: string
          user:
            $ref: '#/definitions/domain.StaffUser'
        type: object
      success:
        default: true
        type: boolean
    type: object
  controller.logoutResponse:
    properties:
      message:
        type: string
      success:
        default: true
        type: boolean
    type: object
...
  controller.refreshResponse:
    properties:
      data:
        properties:
          access_token:
            type: string
        type: object
      success:
        default: true
        type: boolean
    type: object
...
  domain.ErrorDetails:
    properties:
      code:
        type: string
      details:
        type: string
    type: object
  domain.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/domain.ErrorDetails'
      message:
        default: Error
        type: string
      success:
        default: false
        type: boolean
    type: object
...
  domain.StaffUser:
    properties:
      avatar_url:
        type: string
      created_at:
        type: string
      email:
        type: string
      first_name:
        type: string
      id:
        type: string
      last_login_at:
        type: string
      last_name:
        type: string
      role:
        type: string
      status:
        type: string
      updated_at:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /api/admin/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and return JWT access token, and set refresh
        token cookie
      parameters:
      - description: Login credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/controller.loginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.loginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      summary: User Login
      tags:
      - auth
  /api/admin/logout:
    post:
      description: Revoke the user refresh token and clear cookie
      parameters:
      - description: Revoke all sessions/devices
        in: query
        name: all
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.logoutResponse'
      summary: User Logout
      tags:
      - auth
...
  /api/admin/refresh:
    post:
      description: Generate a new access token using the refresh token cookie
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.refreshResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      summary: Refresh Access Token
      tags:
      - auth
...
securityDefinitions:
  BearerAuth:
    description: 'Enter your JWT token as: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"